refactor: reorganize bank account related code by removing unused bank account repository, service, and routes; update imports accordingly
