+++ apps/server/src/users/repository.ts
@@ -1,4 +1,4 @@
-import type { User, CreateUserRequest, UpdateUserRequest, PixHistory, CreateBankAccountRequest, UpdateBankAccountRequest, BankAccount } from './types'
+import type { User, CreateUserRequest, UpdateUserRequest, PixHistory } from './types'
 
 export class UserRepository {
   static async findById(id: string): Promise<User | null> {
@@ -26,28 +26,6 @@ export class UserRepository {
   }
 
   static async addPixHistory(userId: string, pixData: Omit<PixHistory, 'id' | 'userId' | 'createdAt'>): Promise<PixHistory> {
-    throw new Error('Implementar com Prisma')
-  }
-}
-
-export class BankAccountRepository {
-  static async findById(id: string): Promise<BankAccount | null> {
-    return null
-  }
-
-  static async findByConta(conta: string): Promise<BankAccount | null> {
-    return null
-  }
-
-  static async createBankAccount(data: CreateBankAccountRequest): Promise<BankAccount> {
-    throw new Error('Implementar com Prisma')
-  }
-
-  static async updateBankAccount(id: string, data: UpdateBankAccountRequest): Promise<BankAccount | null> {
-    throw new Error('Implementar com Prisma')
-  }
-
-  static async deleteBankAccount(id: string): Promise<boolean> {
     throw new Error('Implementar com Prisma')
   }
 }
