+++ apps/server/src/users/schema.ts
@@ -17,24 +17,3 @@ export const UpdateUserSchema = t.Object({
 export const UserParamsSchema = t.Object({
   id: t.String()
 })
-
-export const CreateBankAccountSchema = t.Object({
-  userId: t.String(),
-  banco: t.String({ minLength: 3 }),
-  agencia: t.String({ minLength: 4 }),
-  conta: t.String({ minLength: 4 }),
-  tipo: t.Enum({ CC: 'CC', CP: 'CP', PP: 'PP' }),
-  tipoPessoa: t.Enum({ F: 'F', J: 'J' })
-})
-
-export const UpdateBankAccountSchema = t.Object({
-  banco: t.Optional(t.String({ minLength: 3 })),
-  agencia: t.Optional(t.String({ minLength: 4 })),
-  conta: t.Optional(t.String({ minLength: 4 })),
-  tipo: t.Optional(t.Enum({ CC: 'CC', CP: 'CP', PP: 'PP' })),
-  tipoPessoa: t.Optional(t.Enum({ F: 'F', J: 'J' }))
-})
-
-export const BankAccountParamsSchema = t.Object({
-  contaBancariaId: t.String()
-})
