+++ apps/server/src/users/types.ts
@@ -33,31 +33,3 @@ export interface UpdateUserRequest {
   pixKey?: string
   cpf?: string
 }
-
-export interface BankAccount {
-  id: UUID
-  userId: UUID
-  banco: string
-  agencia: string
-  conta: string
-  tipo: "CC" | "CP" | "PP"
-  tipoPessoa: "F" | "J"
-  createdAt: Date
-  updatedAt: Date
-}
-
-export interface CreateBankAccountRequest {
-  banco: string
-  agencia: string
-  conta: string
-  tipo: "CC" | "CP" | "PP"
-  tipoPessoa: "F" | "J"
-}
-
-export interface UpdateBankAccountRequest {
-  banco?: string
-  agencia?: string
-  conta?: string
-  tipo?: "CC" | "CP" | "PP"
-  tipoPessoa?: "F" | "J"
-}
